refactor(middleware): Rename the directory name
Correct the spelling.
